Back-End/requirement.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
cachetools==5.3.2
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
conditional==2.0
cron-descriptor==1.4.3
Django==5.0.1
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-clone==5.3.3
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-extensions==3.2.3
django-filter==23.5
django-jazzmin==2.6.0
django-redis==5.4.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
flower==2.0.1
google-auth==2.27.0
google-auth-oauthlib==1.2.0
gspread==6.0.0
gunicorn==22.0.0
hiredis==2.3.2
humanize==4.9.0
idna==3.6
kombu==5.3.6
oauthlib==3.2.2
packaging==24.0
pillow==10.3.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyJWT==2.8.0
python-crontab==3.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
StrEnum==0.4.15
tornado==6.4
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
